id property types

+++ OperationsReporting.Models/DTO/MerchantDto.cs
@@ -4,7 +4,7 @@ namespace OperationsReporting.Models.DTO
 {
     public class MerchantDto
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         
         public required string Name { get; set; }
         
@@ -18,7 +18,7 @@ namespace OperationsReporting.Models.DTO
         
         public string? AddressSecondary { get; set; }
 
-        public int PartnerId { get; set; }
+        public long PartnerId { get; set; }
 
         public List<TransactionDto> Transactions { get; set; } = [];
     }
